Fixed home dashboard (#171)
When we merged PR #162, we introduced a small problem. If the user already had an authentication token stored in browser cache, no dashboard would show up. This had to do with the way HomeDashboard was defined: it was attempting to re-direct the user to either VolunteerDashboard or ShelterDashboard based on the token that's stored in browser cache. However, for now our token does not distinguish between volunteer and shelter admin, so no dashboard would show up instead.